Conclusive overhaul
Switched from a timer for every timed service to a global timer, with an interface to expose the respective callback
Cleaned code up, tried to reduce the amount of global variables to a minimum
