+++ src/Shinoa/Modules/AnimeFeedModule.cs
@@ -1,13 +1,8 @@
 ﻿using Discord;
 using Discord.Commands;
 using Shinoa.Attributes;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
-using System.Xml.Linq;
-using Shinoa.Services;
+using Shinoa.Services.TimedServices;
 
 namespace Shinoa.Modules
 {
