+++ src/Shinoa/Modules/BotAdministrationModule.cs
@@ -7,11 +7,21 @@ using System.Net.Http;
 using System.IO;
 using System.Net;
 using Discord.Commands;
+using Discord.WebSocket;
 
 namespace Shinoa.Modules
 {
     public class BotAdministrationModule : ModuleBase<SocketCommandContext>
     {
+        private readonly DiscordSocketClient client;
+        private readonly CommandService commandService;
+
+        public BotAdministrationModule(DiscordSocketClient clnt, CommandService commandSvc)
+        {
+            client = clnt;
+            commandService = commandSvc;
+        }
+
         [Command("setavatar"), Alias("avatar"), RequireOwner]
         public async Task SetAvatar(string url)
         {
@@ -23,7 +33,7 @@ namespace Shinoa.Modules
             }; 
             var stream = await (await new HttpClient (handler) { BaseAddress = new Uri(splitLink[1].Value) }.GetAsync(splitLink[2].Value))
                 .Content.ReadAsStreamAsync();
-            await Shinoa.DiscordClient.CurrentUser.ModifyAsync(p =>
+            await client.CurrentUser.ModifyAsync(p =>
             {
                 p.Avatar = new Image(stream);
             });
@@ -32,19 +42,19 @@ namespace Shinoa.Modules
         [Command("setplaying"), Alias("setstatus", "game", "status"), RequireOwner]
         public async Task SetPlaying([Remainder]string game)
         {
-            await Shinoa.DiscordClient.SetGameAsync(game);
+            await client.SetGameAsync(game);
         }
 
         [Command("stats"), Alias("diag", "statistics"), RequireOwner]
-        public void GetStats()
+        public async Task GetStats()
         {
-            ReplyAsync(GenerateStatsMessage());
+            await ReplyAsync(GenerateStatsMessage());
         }
 
         [Command("announce"), Alias("announcement", "global"), RequireOwner]
         public async Task Announce([Remainder]string announcement)
         {
-            foreach (IGuild server in Shinoa.DiscordClient.Guilds)
+            foreach (IGuild server in client.Guilds)
             {
                 await (await server.GetDefaultChannelAsync()).SendMessageAsync($"**Announcement:** {announcement}");
             }
@@ -61,7 +71,7 @@ namespace Shinoa.Modules
         [Command("invite"), RequireOwner]
         public async Task GetInviteLink()
         {
-            await ReplyAsync($"Invite link for {Shinoa.DiscordClient.CurrentUser.Mention}: https://discordapp.com/oauth2/authorize?client_id={Shinoa.DiscordClient.CurrentUser.Id}&scope=bot");
+            await ReplyAsync($"Invite link for {client.CurrentUser.Mention}: https://discordapp.com/oauth2/authorize?client_id={client.CurrentUser.Id}&scope=bot&permissions=8");
         }
 
         string GenerateStatsMessage()
@@ -77,7 +87,7 @@ namespace Shinoa.Modules
             output += $"Uptime: {uptimeString}\n\n";
 
             output += "Running modules:\n\n```";
-            foreach (var module in Shinoa.Commands.Modules)
+            foreach (var module in commandService.Modules)
             {
                 output += $"{module.Name}\n";
             }
