+++ src/Shinoa/Services/IService.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections.Generic;
-using Discord.Commands;
+﻿using Discord.Commands;
 
 namespace Shinoa.Services
 {
