+++ src/Shinoa/Services/TimedServices/ITimedService.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace Shinoa.Services.TimedServices
+{
+    public interface ITimedService : IService
+    {
+        Task Callback();
+    }
+}
